feat(formula): split lexer class
